- fixed getting user in I/o thread - fix not found for filter endpoints - fixed log out issue

+++ src/main/kotlin/org/elteq/logic/users/models/Users.kt
@@ -84,7 +84,7 @@ class Users : PanacheEntityBase(), Serializable {
     var firstLogin = true
 
     @Column(name = "token_version")
-    var tokenVersion: Int = 0
+    var tokenVersion: Long = 0
 
     @CreationTimestamp
     @Column(name = "created_on")
